Split charts in different components
